requirements.txt
paddlepaddle>=2.2.0
opencv-python==4.1.2.30
numpy>=1.18.1